+++ src/cli/commands/armor.js
@@ -61,6 +61,15 @@ function configureArmorCommand (armor) {
       return require('./../actions/armor/pull').apply(this, args)
     })
 
+  // dotenvx armor open
+  armor
+    .command('open')
+    .description('open armored key (in browser)')
+    .option('-f, --env-file <path>', 'path to your env file')
+    .action(function (...args) {
+      return require('./../actions/armor/open').apply(this, args)
+    })
+
   // dotenvx armor move
   armor
     .command('move')